Made Group ID 16 bit, added  Call Management protocol doc.

+++ Client/RopuClient.cs
@@ -80,7 +80,7 @@ namespace Ropu.Client
             _retryTimer.Start();
         }
 
-        public void StartCall(uint groupId)
+        public void StartCall(ushort groupId)
         {
             Console.WriteLine("sending StartGroupCall");
             _controllingFunctionClient.StartGroupCall(_userId, groupId);
